add dump-pod http handler

+++ traceloop.go
@@ -19,6 +19,11 @@ var (
 )
 
 func main() {
+	if len(os.Args) == 2 && os.Args[1] == "k8s" {
+		withPidns = true
+		serveHttp = true
+	}
+
 	if len(os.Args) == 2 && os.Args[1] == "guess" {
 		withPidns = true
 	}
@@ -35,7 +40,7 @@ func main() {
 		os.Exit(1)
 	}
 
-	if withPidns {
+	if withPidns && !serveHttp {
 		sig := make(chan os.Signal, 1)
 		signal.Notify(sig, os.Interrupt, os.Kill)
 		select {
@@ -105,6 +110,28 @@ func main() {
 			fmt.Fprintf(w, "%s", out)
 		}
 
+		dumpPodHandler := func(w http.ResponseWriter, r *http.Request) {
+			namespaceStr := r.FormValue("namespace")
+			podnameStr := r.FormValue("podname")
+			idxStr := r.FormValue("idx")
+			if idxStr == "" {
+				fmt.Fprintf(w, "parameter idx missing\n")
+				return
+			}
+			idx, err := strconv.Atoi(idxStr)
+			if err != nil {
+				fmt.Fprintf(w, "%v\n", err)
+				return
+			}
+
+			out, err := t.DumpPod(namespaceStr, podnameStr, idx)
+			if err != nil {
+				fmt.Fprintf(w, "%v\n", err)
+				return
+			}
+			fmt.Fprintf(w, "%s", out)
+		}
+
 		closeHandler := func(w http.ResponseWriter, r *http.Request) {
 			idStr := r.FormValue("id")
 			if idStr == "" {
@@ -137,6 +164,7 @@ func main() {
 		http.HandleFunc("/list", listHandler)
 		http.HandleFunc("/add", addHandler)
 		http.HandleFunc("/dump", dumpHandler)
+		http.HandleFunc("/dump-pod", dumpPodHandler)
 		http.HandleFunc("/dump-by-name", dumpByNameHandler)
 		http.HandleFunc("/dump-by-cgroup", dumpByCgroupHandler)
 		http.HandleFunc("/close", closeHandler)
